refactor(desktop): group download pipeline into main/downloads
