+++ apps/desktop/src/main/downloads/download-queue.test.ts
