+++ apps/desktop/src/main/ipc/downloader.ts
@@ -11,20 +11,20 @@ import {
   getYtDlpVersion,
   getLatestYtDlpVersion,
   downloadYtDlp,
-} from '../ytdlp-manager';
+} from '../downloads/ytdlp-manager';
 import {
   isFFmpegInstalled,
   getFFmpegVersion,
   getLatestFFmpegVersion,
   downloadFFmpeg,
-} from '../ffmpeg-manager';
+} from '../downloads/ffmpeg-manager';
 import { store } from '../app/store';
 import { handle, handleWithFallback } from './with-ipc-handler';
 import { IpcError } from './errors';
 import { invalidate as invalidateFoldersCache } from '../shared/folders-cache';
-import { runYtDlpDownload, type DownloadProgress } from '../yt-dlp-download';
-import { getDownloadQueue } from '../download-queue';
-import { createDownloadQueuePersistence } from '../download-queue-persistence';
+import { runYtDlpDownload, type DownloadProgress } from '../downloads/yt-dlp-download';
+import { getDownloadQueue } from '../downloads/download-queue';
+import { createDownloadQueuePersistence } from '../downloads/download-queue-persistence';
 import type {
   EnqueueDownloadInput,
   ToolInstallResult,
