+++ apps/desktop/src/main/services/metadata-writer.ts
@@ -1,7 +1,7 @@
 import * as path from 'path';
 import * as fs from 'fs';
 import { execFile } from 'child_process';
-import { getFFmpegPath, isFFmpegInstalled } from '../ffmpeg-manager';
+import { getFFmpegPath, isFFmpegInstalled } from '../downloads/ffmpeg-manager';
 import { saveAlbumArt } from '../protocols/art-protocol';
 import { logger } from '../app/logger';
