+++ apps/desktop/src/main/utils/ytdlp-spawn.ts
@@ -1,7 +1,7 @@
 import { spawn } from 'child_process';
 import type { SearchResult } from '@shiranami/contracts';
 import { logger } from '../app/logger';
-import { getYtDlpPath } from '../ytdlp-manager';
+import { getYtDlpPath } from '../downloads/ytdlp-manager';
 import { isHttpUrl } from '../shared/url-safety';
 
 export type { SearchResult };
